First version of the BusinessService
Allows you to specify which groups of Berekeningen belong together and are "in scope" for deriving facts.
Allows you to specify which Glossaries are "in scope" (so the user is not burdened with facts that are not relevant).
Allows you to demand certain facts are filled as input (verplichteInvoer).
Allows you to provide defaults for optional facts (optioneleInvoer).
Allows you to declare the important result facts (uitvoer) to better illustrate the purpose of Berekeningen.

Fixes #82